修改机构管理员的bug

+++ src/main/java/com/wf/controller/AheadUserController.java
@@ -491,21 +491,21 @@ public class AheadUserController {
 		if(StringUtils.isNotBlank(com.getAdminname())&&StringUtils.isNotBlank(com.getAdminpassword())){
 			if(StringUtils.equals(com.getAdminname(), com.getUserId())){
 				hashmap.put("flag", "fail");
-				hashmap.put("fail",  "机构用户的ID和机构管理员的ID重复");
+				hashmap.put("fail",  "机构用户ID和机构管理员ID重复");
 				return hashmap;
 			}
 			Person per=aheadUserService.queryPersonInfo(com.getAdminname());
 			if(per==null){
 				aheadUserService.addRegisterAdmin(com);
 				aheadUserService.addUserAdminIp(com);
-			}else if(per.getUsertype()==1){
+			}else if(per.getUsertype()==1&&StringUtils.equals(per.getInstitution(), com.getInstitution())){
 				//更新机构管理员
 				aheadUserService.deleteUser(com.getAdminname());
 				aheadUserService.addRegisterAdmin(com);
 				//更新管理员IP
 				aheadUserService.deleteUserIp(com.getAdminname());
 				aheadUserService.addUserAdminIp(com);
-			}else if(per.getUsertype()!=1){
+			}else{
 				hashmap.put("flag", "fail");
 				hashmap.put("fail",  "机构管理员的ID已经被占用");
 				return hashmap;
@@ -669,16 +669,11 @@ public class AheadUserController {
 		}
 		if(StringUtils.isNotBlank(com.getAdminname()) || StringUtils.isNotBlank(adminOldName)){
 			if(com.getManagerType().equals("new")){
-				if(StringUtils.equals(com.getAdminname(), com.getUserId())){
-					hashmap.put("flag", "fail");
-					hashmap.put("fail",  "机构用户的ID和机构管理员的ID重复");
-					return hashmap;
-				}
 				Person per=aheadUserService.queryPersonInfo(com.getAdminname());
 				if(per==null){
 					aheadUserService.addRegisterAdmin(com);
 					aheadUserService.addUserAdminIp(com);
-				}else if(per.getUsertype()==1){
+				}else if(per.getUsertype()==1&&StringUtils.equals(per.getInstitution(), com.getInstitution())){
 					//更新机构管理员
 					aheadUserService.deleteUser(com.getAdminname());
 					aheadUserService.addRegisterAdmin(com);
@@ -864,15 +859,10 @@ public class AheadUserController {
 		if(StringUtils.isNotBlank(com.getAdminname()) || StringUtils.isNotBlank(adminOldName)){
 			Person per=aheadUserService.queryPersonInfo(com.getAdminname());
 			if(com.getManagerType().equals("new")){
-				if(StringUtils.equals(com.getAdminname(), com.getUserId())){
-					hashmap.put("flag", "fail");
-					hashmap.put("fail",  "机构用户的ID和机构管理员的ID重复");
-					return hashmap;
-				}
 				if(per==null){
 					aheadUserService.addRegisterAdmin(com);
 					aheadUserService.addUserAdminIp(com);
-				}else if(per.getUsertype()==1){
+				}else if(per.getUsertype()==1&&StringUtils.equals(per.getInstitution(), com.getInstitution())){
 					//更新机构管理员
 					aheadUserService.deleteUser(com.getAdminname());
 					aheadUserService.addRegisterAdmin(com);
@@ -1325,14 +1315,14 @@ public class AheadUserController {
 			if(com.getManagerType().equals("new")){
 				if(StringUtils.equals(com.getAdminname(), com.getUserId())){
 					hashmap.put("flag", "fail");
-					hashmap.put("fail",  "机构用户的ID和机构管理员的ID重复");
+					hashmap.put("fail",  "机构用户ID和机构管理员ID重复");
 					return hashmap;
 				}
 				Person per=aheadUserService.queryPersonInfo(com.getAdminname());
 				if(per==null){
 					aheadUserService.addRegisterAdmin(com);
 					aheadUserService.addUserAdminIp(com);
-				}else if(per.getUsertype()==1){
+				}else if(per.getUsertype()==1&&StringUtils.equals(per.getInstitution(), com.getInstitution())){
 					//更新机构管理员
 					aheadUserService.deleteUser(com.getAdminname());
 					aheadUserService.addRegisterAdmin(com);
